+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import './App.css';
-import Main from './Components/Main/index';
+import Main from './components/Main/index';
 
 function App() {
   return <Main />;
