+++ src/components/LeftBlock/index.js
