+++ src/components/LeftBlock/sidebar/Player.jsx
