+++ src/components/Main/Wall/FormStatus/index.css
